requirements.txt
fastapi==0.104.1
jinja2==3.1.4
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
pydantic==2.5.0
Pillow==11.3.0
numpy>=1.24.0
pymupdf>=1.24.0
python-dotenv==1.0.1
anthropic>=0.40.0
requests>=2.32.0
beautifulsoup4>=4.12.0
